+++ packages/sapient-callout/tsup.config.ts
@@ -2,9 +2,9 @@ import { defineConfig } from 'tsup';
 
 export default defineConfig({
   entry: ['src/index.ts'],
-  format: ['cjs', 'esm'],
-  dts: true,
-  clean: true,
-  external: ['react', 'react-dom'],
+  splitting: false,
   sourcemap: true,
+  clean: true,
+  dts: true,
+  format: ['cjs', 'esm'],
 });