+++ packages/sapient-icons/src/NeutralIcon.tsx
@@ -1,0 +1,18 @@
+/** @jsxImportSource @emotion/react */
+import React from 'react';
+
+export interface IconProps extends React.SVGProps<SVGSVGElement> {
+  size?: number;
+}
+
+export const NeutralIcon = ({ size = 18, ...props }: IconProps) => (
+  <svg width={size} height={size} viewBox="0 0 18 18" fill="none" {...props}>
+    <path
+      d="M9 9L9 13M9 6.03516V6M1 9C1 4.58172 4.58172 1 9 1C13.4183 1 17 4.58172 17 9C17 13.4183 13.4183 17 9 17C4.58172 17 1 13.4183 1 9Z"
+      stroke="currentColor"
+      strokeWidth="1.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+);