+++ packages/sapient-icons/src/SuccessIcon.tsx
@@ -1,0 +1,18 @@
+/** @jsxImportSource @emotion/react */
+import React from 'react';
+
+export interface IconProps extends React.SVGProps<SVGSVGElement> {
+  size?: number;
+}
+
+export const SuccessIcon = ({ size = 18, ...props }: IconProps) => (
+  <svg width={size} height={size} viewBox="0 0 18 18" fill="none" {...props}>
+    <path
+      d="M11.7929 7.20711L8 11L6.70711 9.70711M9 1C4.58172 1 1 4.58172 1 9C1 13.4183 4.58172 17 9 17C13.4183 17 17 13.4183 17 9C17 4.58172 13.4183 1 9 1Z"
+      stroke="currentColor"
+      strokeWidth="1.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+);