+++ packages/sapient-icons/src/index.ts
@@ -1,0 +1,4 @@
+export { NeutralIcon, type IconProps } from './NeutralIcon';
+export { WarningIcon } from './WarningIcon';
+export { ErrorIcon } from './ErrorIcon';
+export { SuccessIcon } from './SuccessIcon';