+++ packages/sapient-design-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sapiently/design-tokens
 
+## 1.0.2
+
+### Patch Changes
+
+- eeyup, party
+
 ## 1.0.1
 
 ### Patch Changes