+++ packages/sapient-primitives/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sapiently/primitives
 
+## 1.0.2
+
+### Patch Changes
+
+- eeyup, party
+- Updated dependencies
+  - @sapiently/design-tokens@1.0.2
+  - @sapiently/theme@1.0.2
+  - @sapiently/core@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes